Modified extraction.c : We can now save the images of the grid and each letters of the grid into the Extraction folder of the project. Case of the non-existent grid and the word list are not coded yet.
Modified extraction.h : All the prototypes of extraction.c functions and the needed #define for them
are written in this file.

Modified Makefile : Added the Extraction folder to the src/ code to be compiled and the -lm flag in
LDFLAGS.

Did not modified the src/Core_Manager.c, as this file was first modified, then rewritten into its
original form.